PM-29876 Fix PR feedback and split StateService and KeychainRepository's server communication config related functions into their own protocols and implementations.
