+++ BitwardenShared/Core/Auth/Services/KeychainRepository.swift
@@ -125,7 +125,7 @@ enum KeychainItem: Equatable, KeychainStorageKeyPossessing {
 
 // MARK: - KeychainRepository
 
-protocol KeychainRepository: AnyObject {
+protocol KeychainRepository: AnyObject, ServerCommunicationConfigKeychainRepository {
     /// Deletes all items stored in the keychain.
     ///
     func deleteAllItems() async throws
@@ -153,12 +153,6 @@ protocol KeychainRepository: AnyObject {
     /// - Parameter userId: The user ID associated with the stored device key.
     ///
     func deletePendingAdminLoginRequest(userId: String) async throws
-
-    /// Attempts to delete the server communication config from the keychain.
-    ///
-    /// - Parameter hostname: The hostname associated with the server communication config.
-    ///
-    func deleteServerCommunicationConfig(hostname: String) async throws
 
     /// Attempts to delete the userAuthKey from the keychain.
     ///
@@ -200,13 +194,6 @@ protocol KeychainRepository: AnyObject {
     /// - Returns: The pending admin login request.
     ///
     func getPendingAdminLoginRequest(userId: String) async throws -> String?
-
-    /// Gets the server communication configuration for a given hostname.
-    ///
-    /// - Parameter hostname: The hostname associated with the server communication config.
-    /// - Returns: The server communication config for that hostname.
-    ///
-    func getServerCommunicationConfig(hostname: String) async throws -> BitwardenSdk.ServerCommunicationConfig?
 
     /// Gets a user auth key value.
     ///
@@ -254,14 +241,6 @@ protocol KeychainRepository: AnyObject {
     ///   - userId: The user ID associated with the pending admin login request.
     ///
     func setPendingAdminLoginRequest(_ value: String, userId: String) async throws
-
-    /// Sets the server communication config for a given hostname.
-    ///
-    /// - Parameters:
-    ///   - config: The server communication config.
-    ///   - hostname: The hostname associated with the config.
-    ///
-    func setServerCommunicationConfig(_ config: BitwardenSdk.ServerCommunicationConfig?, hostname: String) async throws
 
     /// Sets a user auth key/value pair.
     ///
@@ -510,12 +489,6 @@ extension DefaultKeychainRepository {
         )
     }
 
-    func deleteServerCommunicationConfig(hostname: String) async throws {
-        try await keychainService.delete(
-            query: keychainQueryValues(for: .serverCommunicationConfig(hostname: hostname)),
-        )
-    }
-
     func getAccessToken(userId: String) async throws -> String {
         try await getValue(for: .accessToken(userId: userId))
     }
@@ -534,14 +507,6 @@ extension DefaultKeychainRepository {
 
     func getPendingAdminLoginRequest(userId: String) async throws -> String? {
         try await getValue(for: .pendingAdminLoginRequest(userId: userId))
-    }
-
-    func getServerCommunicationConfig(hostname: String) async throws -> BitwardenSdk.ServerCommunicationConfig? {
-        do {
-            return try await getValue(for: .serverCommunicationConfig(hostname: hostname))
-        } catch KeychainServiceError.osStatusError(errSecItemNotFound), KeychainServiceError.keyNotFound {
-            return nil
-        }
     }
 
     func getUserAuthKeyValue(for item: KeychainItem) async throws -> String {
@@ -566,17 +531,6 @@ extension DefaultKeychainRepository {
 
     func setPendingAdminLoginRequest(_ value: String, userId: String) async throws {
         try await setValue(value, for: .pendingAdminLoginRequest(userId: userId))
-    }
-
-    func setServerCommunicationConfig(
-        _ config: BitwardenSdk.ServerCommunicationConfig?,
-        hostname: String,
-    ) async throws {
-        guard let config else {
-            try await deleteServerCommunicationConfig(hostname: hostname)
-            return
-        }
-        try await setValue(config, for: .serverCommunicationConfig(hostname: hostname))
     }
 
     func setUserAuthKey(for item: KeychainItem, value: String) async throws {
