+++ BitwardenShared/Core/Platform/Services/ServiceContainer.swift
@@ -513,7 +513,7 @@ public class ServiceContainer: Services { // swiftlint:disable:this type_body_le
         let sdkRepositoryFactory = DefaultSdkRepositoryFactory(
             cipherDataStore: dataStore,
             errorReporter: errorReporter,
-            stateService: stateService,
+            serverCommunicationConfigStateService: stateService,
         )
         let clientService = DefaultClientService(
             clientBuilder: clientBuilder,
@@ -955,7 +955,7 @@ public class ServiceContainer: Services { // swiftlint:disable:this type_body_le
             sdkRepositoryFactory: DefaultSdkRepositoryFactory(
                 cipherDataStore: dataStore,
                 errorReporter: errorReporter,
-                stateService: stateService,
+                serverCommunicationConfigStateService: stateService,
             ),
             stateService: stateService,
         )
