+++ BitwardenShared/Core/Vault/Repositories/SdkServerCommunicationConfigRepository.swift
@@ -4,32 +4,37 @@ import BitwardenSdk
 final class SdkServerCommunicationConfigRepository: ServerCommunicationConfigRepository {
     // MARK: Properties
 
-    /// The service used by the application to manage account state.
-    private let stateService: StateService
+    /// The service that provides state management functionality for the
+    /// server communication configuration.
+    private let serverCommunicationConfigStateService: ServerCommunicationConfigStateService
 
     // MARK: Init
 
     /// Initializes a `SdkServerCommunicationConfigRepository`
     /// - Parameters:
-    ///   - stateService: The service used by the application to manage account state.
-    init(stateService: StateService) {
-        self.stateService = stateService
+    ///   - serverCommunicationConfigStateService: The service that provides state management functionality for the
+    /// server communication configuration.
+    init(serverCommunicationConfigStateService: ServerCommunicationConfigStateService) {
+        self.serverCommunicationConfigStateService = serverCommunicationConfigStateService
     }
 
     // MARK: Methods
 
     func get(hostname: String) async throws -> BitwardenSdk.ServerCommunicationConfig? {
-        try await stateService.getServerCommunicationConfig(hostname: hostname)
+        try await serverCommunicationConfigStateService.getServerCommunicationConfig(hostname: hostname)
     }
 
     func save(hostname: String, config: BitwardenSdk.ServerCommunicationConfig) async throws {
-        guard let localConfig = try await stateService.getServerCommunicationConfig(hostname: hostname) else {
-            try await stateService.setServerCommunicationConfig(config, hostname: hostname)
+        let localConfig = try await serverCommunicationConfigStateService.getServerCommunicationConfig(
+            hostname: hostname,
+        )
+        guard let localConfig else {
+            try await serverCommunicationConfigStateService.setServerCommunicationConfig(config, hostname: hostname)
             return
         }
 
         let updatedConfig = localConfig.updateCookieValue(from: config)
-        try await stateService.setServerCommunicationConfig(
+        try await serverCommunicationConfigStateService.setServerCommunicationConfig(
             updatedConfig,
             hostname: hostname,
         )
